# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WMS API",
    "version": "1.0.0",
    "description": "Minimal WMS API with Swagger UI"
  },
  "servers": [
    {
      "url": "http://localhost:3100"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/dashboard/storage": {
      "get": {
        "summary": "Get storage totals and breakdown from snapshots",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "YYYY-MM-DD. Defaults to today."
          },
          {
            "in": "query",
            "name": "warehouseId",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "clientId",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "date": {
                      "type": "string",
                      "example": "today"
                    },
                    "filters": {
                      "type": "object",
                      "properties": {
                        "warehouseId": {
                          "type": "integer",
                          "nullable": true
                        },
                        "clientId": {
                          "type": "integer",
                          "nullable": true
                        }
                      }
                    },
                    "totals": {
                      "type": "object",
                      "properties": {
                        "total_cbm": {
                          "type": "number",
                          "example": 120.3456
                        },
                        "total_pallet": {
                          "type": "number",
                          "example": 100.288
                        },
                        "total_sku": {
                          "type": "integer",
                          "example": 350
                        }
                      }
                    },
                    "breakdown": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "warehouse_id": {
                            "type": "integer",
                            "example": 1
                          },
                          "client_id": {
                            "type": "integer",
                            "example": 1
                          },
                          "snapshot_date": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-02-21"
                          },
                          "total_cbm": {
                            "type": "number",
                            "example": 12.3456
                          },
                          "total_pallet": {
                            "type": "number",
                            "example": 10.288
                          },
                          "total_sku": {
                            "type": "integer",
                            "example": 35
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query format"
          }
        }
      }
    },
    "/api/dashboard/storage/snapshots/generate": {
      "post": {
        "summary": "Generate storage snapshots",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "YYYY-MM-DD. Defaults to today."
          },
          {
            "in": "query",
            "name": "warehouseId",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "clientId",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "snapshot_date": {
                          "type": "string",
                          "format": "date",
                          "example": "2026-02-21"
                        },
                        "filters": {
                          "type": "object",
                          "properties": {
                            "warehouseId": {
                              "type": "integer",
                              "nullable": true
                            },
                            "clientId": {
                              "type": "integer",
                              "nullable": true
                            }
                          }
                        },
                        "generated": {
                          "type": "boolean",
                          "example": true
                        },
                        "alerts": {
                          "type": "object",
                          "properties": {
                            "missing_product_cbm_count": {
                              "type": "integer",
                              "example": 0
                            },
                            "missing_product_cbm_items": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "warehouse_id": {
                                    "type": "integer"
                                  },
                                  "client_id": {
                                    "type": "integer"
                                  },
                                  "product_id": {
                                    "type": "integer"
                                  },
                                  "sku_code": {
                                    "type": "string",
                                    "nullable": true
                                  },
                                  "product_name": {
                                    "type": "string"
                                  },
                                  "available_qty": {
                                    "type": "integer"
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query format"
          }
        }
      }
    },
    "/api/dashboard/storage/trend": {
      "get": {
        "summary": "Get storage trend from snapshots",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "groupBy",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "day",
                "week",
                "month"
              ]
            },
            "description": "Default: day"
          },
          {
            "in": "query",
            "name": "warehouseId",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "clientId",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "from": {
                      "type": "string",
                      "format": "date"
                    },
                    "to": {
                      "type": "string",
                      "format": "date"
                    },
                    "groupBy": {
                      "type": "string",
                      "example": "day"
                    },
                    "filters": {
                      "type": "object",
                      "properties": {
                        "warehouseId": {
                          "type": "integer",
                          "nullable": true
                        },
                        "clientId": {
                          "type": "integer",
                          "nullable": true
                        }
                      }
                    },
                    "totals": {
                      "type": "object",
                      "properties": {
                        "total_cbm": {
                          "type": "number",
                          "example": 300.5
                        },
                        "total_pallet": {
                          "type": "number",
                          "example": 250.2
                        },
                        "total_sku": {
                          "type": "integer",
                          "example": 1200
                        }
                      }
                    },
                    "series": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "period": {
                            "type": "string",
                            "example": "2026-02-21"
                          },
                          "total_cbm": {
                            "type": "number",
                            "example": 50.1234
                          },
                          "total_pallet": {
                            "type": "number",
                            "example": 40.4567
                          },
                          "total_sku": {
                            "type": "integer",
                            "example": 200
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query format"
          }
        }
      }
    },
...
  }
}